requirements.txt
groq==0.3.0
google-cloud-texttospeech==2.14.0
google-cloud-speech==2.19.0
google-cloud-translate==3.11.0
pyttsx3==2.90
speechrecognition==3.10.0
pyaudio==0.2.11
rich==13.7.0
python-dotenv==1.0.0

